[rush] Make rush purge also call rush unlink.
